projets/api_rest_108255_2fbe73/requirements.txt
fastapi==0.90.2
uvicorn==0.17.6
motor==2.4.0
pymongo==3.13.0
pydantic==1.10.2
cryptography==39.0.1
python-jose==3.3.0
fastapi-security==0.1.0
fastapi-swagger-ui==0.14.2
fastapi-security-jwks==1.0.0
pytest-pythonpath==0.7.3
pytz==2023.3

[requirements]
fastapi=git+https://github.com/tiangolo/fastapi@v0.90.2#egg=fastapi-0.90.2
uvicorn=git+https://github.com/encode/uvicorn@v0.17.6#egg=uvicorn-0.17.6
python-jose=git+https://github.com/expdev/python-jose@v3.3.0#egg=python-jose-3.3.0

[extra]
